Add backup of data/romajidb.tsv.gz
A pain to remake later, as you need to generate all the JSONL
files first without the RomajiDB (falling back to ambiguous
romaji -> hiragana conversion), then make the RomajiDB, then
regenerate the JSONL. So committing it helps reduce the
bootstrapping required.

As I write this, I realise that compressing the file may not be
better overall, as git will need to store every copy of the gz
instead of just the deltas. Oh well.