Use researchmap name splitter for wikipedia_en
This code can swap names and handles ambiguous romaji for common
names. After comparing the results, 99%+ of the names are now
parsed more correctly (picking the correct first vs. last name
and using long vowels).

There are a handful of counter cases where we guessed
incorrectly:

{"kaki": "周防 正行", "yomi": [-"すお-]{+"すおう+} まさゆき", "authenticity": "real", "lifetime": {"birth_year": 1956, "death_year": null}, "source": "wikipedia_en:Masayuki Suo", }

すお is a valid reading for the surname, but rare enough that we assumed
the longer one. The Wikipedia JA fixes this so one valid solution would
be to ignore entries where we have a JA article

{"kaki": "大植 英次", "yomi": [-"おおうえ-]{+"おおえ+} えいじ", "authenticity": "real", "lifetime": {"birth_year": 1957, "death_year": null}, "source": "wikipedia_en:Eiji Oue", , "notes": "Conductor"}

For this one, the romajidb shows:

大植	oue	sei	おおうえ	おおうえ
大植	oe	sei	おおえ	おおえ

The written name is actually 'Ōue Eiji' so expanding the macron
into 'oo' might help.