NameData: add position, gender fields
With this change we stop (ab)using the tags field for gender and
name position, which are arguably more important than authenticity
but didn't have their own field. We also add an 'is_dict' field for
the 'dict' tag.

Previously tags could contain:
 person (+ masc/fem) for person
 given               for first names
 given  (+ masc/fem) for gendered first names
 surname             for surnames
 unclass             for unclassified names

Additionally surname+masc/fem was supported, and names could have
a combination of types (given+surname+unclass).

This complicated the NameData logic which had to provide is_surname
and is_given_name methods that had to be called in a specific order,
as well as logic to remove tags when changing gender etc.

This change adds two new fields to resolve this, but continues to
support the legacy tags on object creation - this will be fixed
in a later commit. As part of this change, the entire pipeline was
tested again, which unearthed new issues:

 1) subreadings were not being imported in some cases
 2) hits_unknown was not being considered for kanji_stats, which
    led to the stats being incorrect
 3) names such as 一の瀬 (containing kana) were being rejected
 4) jmnedict entries containing katakan were being rejected
 5) ignore list was missing some useful items such as 大学

All are now resolved.